# openapi.yaml
openapi: 3.0.3
info:
  title: QR Reward Platform API
  version: '1.0'
  description: |
    Promoters fund QR campaigns; end users scan and sign up with a publisher; coins move
    automatically, exactly once. See `SYSTEM_FLOW.md` for the end-to-end narrative behind
    these endpoints and the invariants each one enforces.

    ### Auth schemes
    - **sessionAuth** — `Authorization: Bearer <jwt>` from `/v1/auth/login` or `/v1/auth/signup`.
      12h expiry. Used by promoter/publisher portal routes and the admin console.
    - **apiKeyAuth** — `Authorization: Bearer pk_<hex>` issued once at publisher signup (or via
      rotate). Used only by the publisher's own server, never a browser.
    - Unauthenticated routes: `/healthz`, `GET /r/:code`, `GET /v1/qr-codes/:id/image`,
      `POST /v1/auth/signup`, `POST /v1/auth/login`.
servers:
  - url: http://localhost:4000
    description: Local dev
  - url: '{baseUrl}'
    description: Deployed
    variables:
      baseUrl:
        default: https://api.example.com
security: []
...
paths:
...
  /v1/redemptions:
    get:
      tags: [Portal]
      summary: My org's recent redemptions (portal view)
      description: Latest 100, either side of the partnership.
      security: [{ sessionAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Redemption' } }

  # ---------------- Partner API ----------------
  /v1/redemptions/verify:
    post:
      tags: [Partner API]
      summary: Grant a reward for a scanned-then-signed-up user
      description: |
        Called by the publisher's own server after a user signs up, carrying the scan token
        from the landing-page redirect. Pays `coin_rate` if `identified: true`, else
        `guest_rate` (the default) with the remainder held for `grace_days`. One redemption
        per `(campaign, publisher_user_ref)` — a second call for the same user conflicts.
      security: [{ apiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scan_token, publisher_user_ref]
              properties:
                scan_token: { type: string, description: The `st` query param from the landing redirect }
                publisher_user_ref: { type: string, description: The publisher's own user id }
                identified:
                  type: boolean
                  default: false
                  description: Publisher's own verification bar cleared. Absent/false = guest tier.
      responses:
        '201':
          description: Reward granted
          content:
            application/json:
              schema:
                type: object
                properties:
                  redemption_id: { type: string, format: uuid }
                  campaign_id: { type: string, format: uuid }
                  coins: { type: integer }
                  identified: { type: boolean }
                  status: { type: string, example: granted }
                  pending_coins: { type: integer, description: What upgrading later would still pay }
                  upgrade_deadline: { type: string, format: date-time, nullable: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: budget_exhausted | scan token already used | duplicate_user
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '422':
          description: scan token invalid/expired, or not a scan token
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /v1/redemptions/{id}/upgrade:
    post:
      tags: [Partner API]
      summary: Release the held-back guest→verified delta
      description: |
        Idempotent — calling it twice on an already-upgraded redemption is a no-op
        (`status: already_full`). Refused once `grace_days` has elapsed since the original scan.
      security: [{ apiKeyAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '201':
          description: Upgraded (or already full)
          content:
            application/json:
              schema:
                type: object
                properties:
                  redemption_id: { type: string, format: uuid }
                  coins: { type: integer }
                  coins_added: { type: integer }
                  identified: { type: boolean }
                  status: { type: string, enum: [upgraded, already_full] }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: budget_exhausted | grace_period_expired
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /v1/redemptions/{id}:
    get:
      tags: [Partner API]
      summary: Look up a redemption
      security: [{ apiKeyAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/Redemption' } } }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------- Admin ----------------
...
components:
  securitySchemes:
    sessionAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 12h session token from `/v1/auth/signup` or `/v1/auth/login`
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: 'pk_<hex>'
      description: Publisher's Partner API key — server-to-server only
  responses:
...
    Unauthorized:
      description: Missing/invalid credentials, or account suspended
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    NotFound:
      description: Not found (or not owned by the caller — ownership failures read as 404, not 403)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      properties:
        statusCode: { type: integer }
        message: { type: string }
        error: { type: string }
...
    Redemption:
      type: object
      properties:
        id: { type: string, format: uuid }
        campaign_id: { type: string, format: uuid }
        campaign_name: { type: string }
        scan_id: { type: string, format: uuid }
        publisher_user_ref: { type: string }
        coins: { type: integer }
        identified: { type: boolean }
        upgraded_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
...